Implement training for sigma network

+++ convex_network/net_maker.py
@@ -16,12 +16,12 @@ def get_trained_net_and_test_set(df: pd.DataFrame, test_df: pd.DataFrame, test_s
         if not analytics_mode:
             df = df[df['avg_type'] == OptionAvgType.ARITHMETIC.value]
         df_values = df[['spot_strike_ratio', 'ttm', 'risk_free_rate', 'volatility']].astype(
-            np.float32).to_numpy()
+            np.float64).to_numpy()
     elif fixed_avg_type == OptionAvgType.GEOMETRIC:
         if not analytics_mode:
             df = df[df['avg_type'] == OptionAvgType.GEOMETRIC.value]
         df_values = df[['spot_strike_ratio', 'ttm', 'risk_free_rate', 'volatility']].astype(
-            np.float32).to_numpy()
+            np.float64).to_numpy()
     else:
         if not analytics_mode:
             df['numeric_avg_type'] = df.apply(lambda row: 1 if row.avg_type == OptionAvgType.ARITHMETIC.value else 0,
@@ -31,20 +31,20 @@ def get_trained_net_and_test_set(df: pd.DataFrame, test_df: pd.DataFrame, test_s
                 axis=1)
         if WITH_CI_STATS:
             df_values = df[['spot_strike_ratio', 'ttm', 'risk_free_rate', 'volatility', 'numeric_avg_type', 'left_ci', 'right_ci']].astype(
-                np.float32).to_numpy()
+                np.float64).to_numpy()
             test_df_values = test_df[
                 ['spot_strike_ratio', 'ttm', 'risk_free_rate', 'volatility', 'numeric_avg_type', 'left_ci',
                  'right_ci']].astype(
-                np.float32).to_numpy()
+                np.float64).to_numpy()
         else:
             df_values = df[['spot_strike_ratio', 'ttm', 'risk_free_rate', 'volatility', 'numeric_avg_type']].astype(
-                np.float32).to_numpy()
+                np.float64).to_numpy()
             test_df_values = test_df[
                 ['spot_strike_ratio', 'ttm', 'risk_free_rate', 'volatility', 'numeric_avg_type']].astype(
-                np.float32).to_numpy()
+                np.float64).to_numpy()
 
-    df_target = df['price_strike_ratio'].astype(np.float32).to_numpy()
-    test_df_target = test_df['price_strike_ratio'].astype(np.float32).to_numpy()
+    df_target = df['price_strike_ratio'].astype(np.float64).to_numpy()
+    test_df_target = test_df['price_strike_ratio'].astype(np.float64).to_numpy()
 
     if test_size == 1 and USE_PRETRAINED_NET:
         x_test = df_values
